Add possibility to upVote and downVote an opinion
There is a bug : The feature says that a user can vote without having commented and a user can already have multiple comments in an opinion.
The database is not designed for this, whether a user can only have one comment in an opinion, or the DB need a `vote` table in addition to the existing `user_opinion` table.

On hold, waiting for more information.